refactor(test): display name 변경
