+++ src/test/java/nextstep/subway/favorite/FavoriteAcceptanceTest.java
@@ -3,6 +3,6 @@ package nextstep.subway.favorite;
 import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 
-@DisplayName("즐겨찾기 관련 기능")
+@DisplayName("즐겨찾기 관련 기능 인수 테스트")
 public class FavoriteAcceptanceTest extends AcceptanceTest {
 }
